DefaultAutoCommand complete!!
Fixed and made it a lot easier to change and work with the setVelocityTarget() method. Got rid of the strings and the weird process of choosing the autos. I also made the DefaultAutoCommand and put the code previously from RobotContainer and I pasted it into there (and I also fixed so the robot wont travel diagonally, but rather straight backwards).